add v2.0.0 pipeline build

+++ inst/scripts/make-metadata.R
@@ -1,10 +1,13 @@
 make_metadata <-
 function(
-    directory = "~/github/MultiAssayExperiment.TCGA/",
-    dataDir = "data/bits",
-    ext_pattern = allextpat,
+    directory,
+    dataDir,
+    version,
+    ext_pattern,
     resource_maintainer = utils::maintainer("curatedTCGAData"),
-    resource_biocVersion = BiocManager::version())
+    resource_biocVersion = BiocManager::version(),
+    append = FALSE
+)
 {
     exdata <- "inst/extdata"
     metafile <- file.path(exdata, "metadata.csv")
@@ -12,12 +15,16 @@ function(
     if (!dir.exists(exdata))
         dir.create(exdata)
 
-    if (file.exists(metafile))
+    if (file.exists(metafile) && !append)
         file.remove(metafile)
 
     metadat <- getMetadata(directory = directory, dataDir = dataDir,
-        ext_pattern = ext_pattern, resource_maintainer = resource_maintainer,
+        version = version, ext_pattern = ext_pattern,
+        resource_maintainer = resource_maintainer,
         resource_biocVersion = resource_biocVersion)
+
+    if (append)
+        metadat <- rbind(readr::read_csv(metafile), metadat)
 
     readr::write_csv(metadat, "inst/extdata/metadata.csv", col_names = TRUE)
 }
